+++ CHANGELOG.md
@@ -117,7 +117,7 @@
 
 ## 4.17.0-dev.1
 
-* feat: add static code diagnostic [`avoid-use-expanded-as-spacer`](https://dartcodemetrics.dev/docs/rules/flutter/avoid-use-expanded-as-spacer).
+* feat: add static code diagnostic [`avoid-expanded-as-spacer`](https://dartcodemetrics.dev/docs/rules/flutter/avoid-expanded-as-spacer).
 * feat: migrate to new analyzer plugins API.
 * chore: changed min `SDK` version to `2.17.0`.
 * chore: restrict `analyzer` version to `>=4.1.0 <4.3.0`.